refactor utils and edit docstrings
