+++ client/blindai/utils/utils.py
@@ -15,7 +15,6 @@
 import re
 import os
 import cryptography.x509
-import grpc
 from cryptography.hazmat.primitives.serialization import Encoding
 from cryptography.x509 import ObjectIdentifier, load_pem_x509_certificate
 from cryptography.hazmat.primitives.asymmetric.ed25519 import Ed25519PublicKey
@@ -40,33 +39,6 @@ def create_byte_chunk(data):
         sent_bytes += min(CHUNK_SIZE, len(data) - sent_bytes)
 
 
-def check_rpc_exception(rpc_error):
-    if rpc_error.code() == grpc.StatusCode.CANCELLED:
-        return f"Cancelled GRPC call: code={rpc_error.code()} message={rpc_error.details()}"
-
-    elif rpc_error.code() == grpc.StatusCode.UNAVAILABLE:
-        return f"Failed to connect to GRPC server: code={rpc_error.code()} message={rpc_error.details()}"
-
-    else:
-        return (
-            f"Received RPC error: code={rpc_error.code()} message={rpc_error.details()}"
-        )
-
-
-def check_socket_exception(socket_error):
-    if len(socket_error.args) >= 2:
-        error_code = socket_error.args[0]
-        error_message = socket_error.args[1]
-        return f"Failed To connect to the server due to Socket error : code={error_code} message={error_message}"
-
-    elif len(socket_error.args) == 1:
-        error_message = socket_error.args[0]
-        return f"Failed To connect to the server due to Socket error : message={error_message}"
-
-    else:
-        return f"Failed To connect to the server due to Socket error "
-
-
 def get_enclave_signing_key(server_cert):
     ENCLAVE_ED25519_SIGNING_KEY_OID = ObjectIdentifier("1.3.6.1.3.2")
     enclave_ed25519_signing_key = (
@@ -78,12 +50,6 @@ def get_enclave_signing_key(server_cert):
         enclave_ed25519_signing_key
     )
     return enclave_signing_key
-
-
-class SignatureError(Exception):
-    """This exception is raised when the signature or the returned digest is invalid"""
-
-    pass
 
 
 def read(filename):
